refactored for running coffeescript tests, and added #458 test in coffeescript
