+++ test/coffee/run.coffee
@@ -1,0 +1,3 @@
+require "../sinon/issues/issues.coffee"
+
+require "../run-config.js"
