+++ test/node/run.js
@@ -16,31 +16,5 @@ require("../sinon/typeOf_test.js");
 require("../sinon/log_error_test.js");
 require("../sinon/issues/issues.js");
 
-var buster = require("../runner");
-var args = process.argv.slice(2);
 
-while (args.length) {
-    switch (args.shift()) {
-        case "-h":
-        case "--help":
-            console.log("Usage: node test/node/run.js [-u]");
-            console.log("  -u, --unstable  Enable unstable tests");
-            console.log("  -h, --help      This cruft");
-            process.exit(1);
-            break;
-        case "-u":
-        case "--unstable":
-            require("../sinon/util/fake_timers_test.js");
-            break;
-    }
-}
-
-buster.testRunner.onCreate(function (runner) {
-    runner.on("suite:end", function (results) {
-        // Reporter will be set up after delay, allow
-        // it to finish before we exit the process
-        process.nextTick(function () {
-            process.exit(results.ok ? 0 : 1);
-        });
-    });
-});
+require("../run-config.js");
